refactor: 타입 변경 및 추가사항 변경
