+++ src/main/java/com/d108/project/domain/forum/board/domain/Board.java
@@ -1,4 +1,4 @@
-package com.d108.project.domain.board;
+package com.d108.project.domain.forum.board.domain;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -7,16 +7,18 @@ import lombok.Getter;
 @Getter
 @Table(name = "boards")
 public class Board {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "board_id")
-    private Long id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
 
     // FIXME: 상권 (하나의 상권에 해당하는 하나의 게시판이 생기는 거 아님?)
-
+    private Integer businessAreaId;
     // FIXME: 프랜차이즈 (이하 동일)
 
+    private Integer franchiseId;
+
     // 제목
-    private String subject;
+    private String title;
 
     // 내용
     private String content;
